#123: fix div issue and move views to pages
