+++ frontend/src/components/pages/namespace/NamespaceView.tsx
@@ -3,34 +3,24 @@
 // SPDX-FileCopyrightText: 2019 Georg Schwarz <georg. schwarz@fau.de>
 
 import React, { useState } from 'react'
-import ProducerModal from '../modals/ProducerModal'
-import { Collapse, CardActions, Button } from '@mui/material'
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
 import ExpandLessIcon from '@mui/icons-material/ExpandLess'
+import { Collapse, CardActions, Button } from '@mui/material'
 import { useNavigate } from 'react-router-dom'
-import { addFilter } from '../../store/filterSlice'
-import { useAppDispatch } from '../../store/hooks'
+import { addFilter } from '../../../store/filterSlice'
+import { useAppDispatch } from '../../../store/hooks'
 
-const TopicView: React.FC<TopicViewProps> = ({ data }) => {
-	const { name, tenant, namespace }: TopicInfo = data
-
-	/*
-	const topicConsumers = data?.topicStatsDto?.subscriptions
-		.map((item: SampleSubscription) => item.consumers)
-		.filter((el: Array<string>) => el && el.length > 0)
-		.flat()
-	*/
-	//const topicProducers = data?.topicStatsDto?.producers
+const NamespaceView: React.FC<NamespaceViewProps> = ({ data }) => {
+	const { id, tenant }: NamespaceInfo = data
 
 	const [expanded, setExpanded] = useState(false)
 	const dispatch = useAppDispatch()
 	const navigate = useNavigate()
 
 	const handleDrillDown = () => {
-		dispatch(addFilter({ filterName: 'topic', id: name }))
-		navigate('/message')
+		dispatch(addFilter({ filterName: 'namespace', id: id }))
+		navigate('/topic')
 	}
-
 	const handleExpand = () => {
 		//TODO if(!data) fetch detailed data
 		setExpanded(!expanded)
@@ -38,17 +28,13 @@ const TopicView: React.FC<TopicViewProps> = ({ data }) => {
 
 	return (
 		<div className="flex flex-col card-content">
-			<h2>{name}</h2>
+			<h2 className="uppercase">{id}</h2>
 			<div className="flex card-inner">
 				<div className="flex flex-col card-col card-col-1">
 					<div className="flex flex-col card-info">
 						<p className="text-black">
 							Tenant:{' '}
 							<span className="text-blue">{tenant ? tenant : 'N/A'}</span>
-						</p>
-						<p className="text-black">
-							Namespace:{' '}
-							<span className="text-blue">{namespace ? namespace : 'N/A'}</span>
 						</p>
 					</div>
 				</div>
@@ -59,68 +45,52 @@ const TopicView: React.FC<TopicViewProps> = ({ data }) => {
 					<div className="flex flex-col card-col card-col-1">
 						{/*<div className="flex flex-col card-info">
 							<p className="text-black">
-								Producers:{' '}
+								Bundles:{' '}
 								<span className="text-blue">
-									{topicProducers &&
-										topicProducers.length > 0 &&
-										topicProducers.map((item: string) => (
-											<ProducerModal
-												key={'producer-' + Math.floor(Math.random() * 999999)}
-												producer={{
-													producerName: item,
-													topicList: ['SampleTopic1', 'SampleTopic2'],
-													messageList: ['SampleMessage1', 'SampleMessage2'],
-												}}
-											/>
-										))}
+									{data?.bundlesData?.numBundles
+										? data.bundlesData?.numBundles
+										: 0}
 								</span>
 							</p>
 							<p className="text-black">
-								Consumers:{' '}
+								Message TTL:{' '}
 								<span className="text-blue">
-									{/*topicConsumers &&
-									topicConsumers.length > 0 &&
-									topicConsumers.map((item: string, index: number) => (
-										<ConsumerModal
-											key={'consumer-' + Math.floor(Math.random() * 999999)}
-											consumer={{
-												consumerName: item,
-												topicList: ['SampleTopic1', 'SampleTopic2'],
-												messageList: ['SampleMessage1', 'SampleMessage2'],
-											}}
-										/>
-										))
+									{data?.messagesTTL ? data?.messagesTTL : 'None'}
+								</span>
+							</p>
+							<p className="text-black">
+								Retention time:{' '}
+								<span className="text-blue">
+									{data?.retentionPolicies?.retentionTimeInMinutes
+										? data?.retentionPolicies.retentionTimeInMinutes +
+										  ' minutes'
+										: 'None'}
+								</span>
+							</p>
+							<p className="text-black">
+								Retention size:{' '}
+								<span className="text-blue">
+									{data?.retentionPolicies?.retentionSizeInMB
+										? data?.retentionPolicies.retentionSizeInMB + ' MB'
+										: 'None'}
 								</span>
 							</p>
 						</div>
 						<div className="grey-line"></div>
 						<div className="flex flex-col card-info">
 							<p className="text-black">
-								Produced messages:{' '}
+								Cluster:{' '}
 								<span className="text-blue">
-									{data?.topicStatsDto?.producedMesages
-										? data.topicStatsDto.producedMesages
-										: 0}
+									{data?.cluster ? data.cluster : 'N/A'}
 								</span>
 							</p>
 							<p className="text-black">
-								Average message size:{' '}
+								Topics:{' '}
 								<span className="text-blue">
-									{data?.topicStatsDto?.averageMessageSize
-										? data.topicStatsDto.averageMessageSize
-										: 0}
+									{data?.amountOfTopics ? data.amountOfTopics : 0}
 								</span>
 							</p>
-							<p className="text-black">
-								Storage size:{' '}
-								<span className="text-blue">
-									{data?.topicStatsDto?.storageSize
-										? data.topicStatsDto.storageSize
-										: 0}
-								</span>
-							</p>
-						</div>
-						*/}
+						</div>*/}
 					</div>
 				</div>
 			</Collapse>
@@ -155,4 +125,4 @@ const TopicView: React.FC<TopicViewProps> = ({ data }) => {
 	)
 }
 
-export default TopicView
+export default NamespaceView
